==> pipelines/13_psns_comparative_advantages_extraction.ipynb <==
# IMPORTANT: The parameters below are set only for running this notebook independently. 
# When executing the full Ploomber pipeline, these values will be overridden by the settings in `pipeline.yaml`. 
# Any modifications made here will not persist when running the pipeline.
upstream = None
COUNTRY =  'ARM' # Code of the Country
product = {'data': f'../data/processed/{COUNTRY}/psns_comparative_advantages.xlsx'}  # Path to save the final data product (stored under the 'data' key)
data_source = 'data/corpora/psns/'  # Path to the source data directory

This Notebook uses an LLM model to analyze the Programme Strategy Notes to:
-	summarize sections and identify key UNICEF’s comparative advantages. 
-	summarize national priorities for all countries. These priorities are categorized into one or more goal areas.

import re
from pathlib import Path
import pandas as pd
import unicef_cpe 
from unicef_cpe.config import PROJ_ROOT


MODEL = 'gpt-4o'
country_map = {k:v for k,v in unicef_cpe.utils.get_ecaro_countries_mapping(priority=False).items() if k in COUNTRY}
country_code_map = {v:k for k,v in country_map.items() }

### Extract Reccomendations from UNSDCF

data = []
source_path =  PROJ_ROOT / Path(data_source)

for file_path in source_path.glob('*.pdf'):
    with open(file_path, 'r') as file:
        country_code, *_ = file_path.name.split('-')
        country_code = country_code.upper()
        country = country_map.get(COUNTRY)
        year = re.search(r'\d{4}', file_path.name).group()
        text = unicef_cpe.extraction.extract_text_from_file(file_path)
        data.append((file_path.name, country_code, country, year, text))
df_doc = pd.DataFrame(data, columns=['file_name', 'country_code','country', 'year', 'section_text'])
df_doc.sort_values(['country_code', 'country', 'year', 'file_name'], ignore_index=True, inplace=True)
print('Shape:', df_doc.shape)
df_doc = df_doc[df_doc['country_code'].eq(COUNTRY)]
print(df_doc.head())


data = []

# Iterate over each document in the original DataFrame
for index, row in df_doc.iterrows():
    year = row['year']
    
    # Extract ection titles
    section_titles = unicef_cpe.document_processing.extract_section_titles(row['section_text'], threshold=50, base_min_lines_between_titles=10, start_with_digit=False)

    # Filter only Section from First Chapter. This is the relevant one
    filtered_section_titles = [x for x in section_titles if 'advantage' in x.lower()]

    if len(filtered_section_titles)==0 and 'advantage' in row['section_text'].lower():
        # Extract chucks of text that contain "advantage" in their text
        sections = unicef_cpe.document_processing.extract_keyword_chunks(row['section_text'], 'advantage', token_window=50)

    else:
        # Extract sections and their text based on the section titles
        sections = unicef_cpe.document_processing.extract_sections_with_text(row['section_text'], filtered_section_titles)

    # Add each section and its text to the new DataFrame
    for section_title, section_text in sections:
        # Clean the text 
        cleaned_section_text = unicef_cpe.utils.clean_text(section_text)

        subjects = unicef_cpe.genai.summarise_comparative_advantage(section_text, model=MODEL, api_type='openai')

        data.append({
            'country_code': row['country_code'],
            'country': row['country'],
            'section_title': section_title,
            'section_text': cleaned_section_text,
            'comparative_advantage': subjects,
            'year': row['year'],
            'file_name': row['file_name'],
        })

# Create a new DataFrame from the collected data
df_sections = pd.DataFrame(data)

df_sections = df_sections.groupby(['country_code', 'country', 'year', 'file_name']).agg({
    'section_title': '\n'.join,
    'section_text': '\n'.join,
    'comparative_advantage': '\n'.join
}).reset_index()

df_list = []

country_name = country_map.get(COUNTRY)


country_mask = df_sections['country'] == country_name

years = sorted(df_sections[country_mask]['year'].unique())

for year in years:
    year_mask = df_sections['year'] == year
    
    try:
        df_filtered = df_sections[country_mask & year_mask].copy()
        df_filtered['title_text_combined'] = df_filtered['section_title'] + '\n' + df_filtered['section_text']

        # Join all rows by '\n\n'
        text = '\n\n'.join(df_filtered['title_text_combined'].tolist())

    except IndexError:
        # Skip if there's no narrative text for this combination of country and year
        print(f"No narrative text available for {country_name} in {year}")
        continue
    print(f"Summarizing national priorities for {country_name} in {year}")
    # Extract needs and categorize them by goal area
    subjects = unicef_cpe.genai.extract_by_goal_area(text, subject = "national priorities", model=MODEL, api_type='openai')

    subjects ='\n'+ subjects
    subject_list = subjects.split('\n- **')[1:]  # Skip the first empty element after the split
    for subject in subject_list:
        # Split each need into goal area and its description
        subject_description = subject.split('**:')
        if len(subject_description)>1:
            goal_area = subject_description[0].replace('- **','').replace('**', '').replace('\n', '').strip()
            new_subject = subject_description[1].replace('\n', '').strip()
            df_list.append([COUNTRY, country_name, year, goal_area, new_subject])

df_advantage = pd.DataFrame(df_list, columns=['country_code', 'country', 'year', 'goal_area', 'comparative_advantage'])

print(df_advantage.head())

output_path = Path(product['data'])
output_path.parent.mkdir(parents=True, exist_ok=True)  # Create missing directories
df_advantage.to_excel(product['data'], index=False)

################################################################################################################################################################################################
